Fix NaN gradients in ranking_loss + more robust training loop
ranking_loss had a dtype mismatch: y_pred is float64 (model input dtype)
but t/event/t_last were cast to float32. The mixed-precision subtraction
inside _log_S (float32 t_val - float64 log_alpha) produced NaN gradients
starting around epoch 3 once the model reached a critical weight state.
Fix: cast alpha/beta to float32 at the start of ranking_loss so all
intermediate ops stay in one dtype.

training_loop changes:
- NaN/Inf guard on combined loss before tape.gradient (skips the batch
  immediately rather than computing NaN gradients first)
- Switched per-tensor clip_by_norm to clip_by_global_norm: keeps all
  gradient tensors proportional while bounding the total update norm,
  which is more effective when ranking_loss and weibull_loss contribute
  at different scales across layers
- ranking_loss NaN/Inf guard: skips adding rl to loss if it is non-finite

config_stklm0.yaml: gamma 0.5 → 0.1 (0.5 was too large for 0.6% CSM
event rate; even with stratified batching the ranking loss gradient
dominated and caused instability)

+++ bertpca/src/bertpca/loss.py
@@ -166,11 +166,12 @@ def ranking_loss(y_true: tf.Tensor, y_pred: tf.Tensor) -> tf.Tensor:
     event  = tf.cast(y_true[:, 1], tf.float32)
     t_last = tf.cast(y_true[:, 2], tf.float32)
 
-    # Same guards as weibull_loss: tf.maximum prevents alpha/beta going
-    # negative (raw model output can be arbitrarily negative early in training,
-    # and x**non_integer for x<0 is NaN).
-    alpha = tf.maximum(y_pred[:, 0] + 1.0, EPSILON)
-    beta  = tf.maximum(y_pred[:, 1] + 1.0, EPSILON)
+    # Cast y_pred to float32 so all computations stay in one dtype.
+    # y_pred comes out of the model as float64 (input dtype=float64), but
+    # t/event/t_last above are float32.  Mixed-precision intermediate ops
+    # (float32 - float64 subtraction inside _log_S) can produce NaN gradients.
+    alpha = tf.maximum(tf.cast(y_pred[:, 0], tf.float32) + 1.0, EPSILON)
+    beta  = tf.maximum(tf.cast(y_pred[:, 1], tf.float32) + 1.0, EPSILON)
 
     def _log_S(t_val, a, b):
         # log-survival: -(t/a)^b in log-space to prevent float32 overflow
